Move FitsOpener into settings
