+++ stellarphot/gui_tools/__init__.py
@@ -1,6 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
 from .comparison_functions import *
-from .fits_opener import *
 from .photometry_widget_functions import *
 from .seeing_profile_functions import *
