+++ stellarphot/settings/fits_opener.py
